Move Profile-specific code into vap-svc dir
- update the showAddressValidationModal code comment
