+++ src/platform/user/profile/vap-svc/tests/addressValidationMessages.unit.spec.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { expect } from 'chai';
-import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import sinon from 'sinon';
 import {
   ADDRESS_VALIDATION_MESSAGES,
   ADDRESS_VALIDATION_TYPES,
-} from '../../profile/constants/addressValidationMessages';
+} from '../constants/addressValidationMessages';
 
 describe('ADDRESS_VALIDATION_MESSAGES object', () => {
-  it('has a valid data object for each validation type', () => {
+  it('has a valid data object for each ADDRESS_VALIDATION_TYPE', () => {
     const validationTypes = Object.values(ADDRESS_VALIDATION_TYPES);
     validationTypes.forEach(type => {
       const messageData = ADDRESS_VALIDATION_MESSAGES[type];
@@ -16,10 +16,10 @@ describe('ADDRESS_VALIDATION_MESSAGES object', () => {
       expect(typeof messageData).to.equal('object');
       expect(typeof messageData.headline).to.equal('string');
       expect(typeof messageData.ModalText).to.equal('function');
-      const modalTextComponent = shallow(
+      const modalTextComponent = mount(
         <messageData.ModalText editFunction={editSpy} />,
       );
-      modalTextComponent.find('a').simulate('click');
+      modalTextComponent.find('button').simulate('click');
       expect(editSpy.called).to.be.true;
       modalTextComponent.unmount();
     });
